+++ modules/instances/variables.tf
@@ -63,6 +63,8 @@ variable "instances" {
     volume_size              = number
     state                    = string
     autherized_keys          = string
+    ipv6                     = optional(bool, false)
+    defined_tags             = optional(map(string), null)
     config = object({
       shape             = string
       flex_shape_config = map(string)
@@ -91,11 +93,19 @@ variable "instances" {
       nsg_ids                = list(string)
       skip_source_dest_check = bool
       hostname_label         = string
+      ipv6                   = optional(bool, false)
       secondary_ips = map(object({
         name       = string
         ip_address = string
       }))
     }))
+    agent_plugins = optional(
+      map(object({
+        name       = string
+        is_enabled = bool
+      })),
+      {}
+    )
     optionals = map(any)
     # preserve_boot_volume =  bool (true)
     # boot_volume_id = string